InfluxDB/requirements.txt
influxdb-client==1.38.0
requests==2.31.0
pandas>=2.0.0
numpy>=1.24.0
kafka-python==2.0.2
